[Close #228] Last PostController action class.
Pinging Issue #54.

RSpec: 688 examples, 0 failures
RuboCop: 229 files inspected, no offenses detected
