--- lib/newpoc-action-post-update/.rspec
@@ -1,2 +1,0 @@
---format documentation
---color
