--- lib/newpoc-action-post-update/.travis.yml
@@ -1,3 +1,0 @@
-language: ruby
-rvm:
-  - 2.1.5
