--- lib/newpoc-action-post-update/lib/newpoc/action/post/update/version.rb
@@ -1,11 +1,0 @@
-
-module Newpoc
-  module Action
-    module Post
-      # Verifies that the current user is permitted to update a specified post.
-      class Update
-        VERSION = '0.0.2'
-      end
-    end
-  end
-end
