--- lib/newpoc-action-post-update/spec/spec_helper.rb
@@ -1,13 +1,0 @@
-
-require 'simplecov'
-SimpleCov.start do # 'rails'
-  add_filter '/spec/'
-  add_filter '/vendor/'
-end
-
-$LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
-
-require 'awesome_print'
-require 'pry'
-
-require 'newpoc/action/post/update'
